changed so that divide returns double

+++ Lektion 2/src/Math.java
@@ -8,7 +8,7 @@ public class Math {
         return a * b;
     }
 
-    public int divide(int a, int b) {
+    public double divide(double a, double b) {
         return a / b;
     }
